+++ .pm/tasks/task_f81d3e661b7048d8b2fe6987a544a368.execution.md
@@ -279,3 +279,12 @@ Example:
 - Finding Disposition Evidence: `filter_request_peers_by_health` now preserves record-exchange-pending soft blocked peers as fallback candidates while excluding empty/hard blocked peers; regression `filter_request_peers_by_health_keeps_record_exchange_pending_blocked_peer_as_fallback`; 57 `oasis7_net` libp2p tests and 8 replication peer selection tests passed
 - Residual Risk: End-to-end proof still depends on merging this follow-up, rebuilding CI package from the new merge commit, and validating live local/Windows public testnet high-state sync against real bootstrap timing.
 - Blocker / Next Action: Amend commit with review evidence, run PR preflight/create, watch checks, merge, package, deploy.
+
+## 2026-06-12 01:16:58 CST / tpm
+- 完成内容: Addressed PR #412 required-gate file-size blocker without changing production behavior.
+- 遗留事项: GitHub checks need to rerun on the pushed test module split; merge, CI package, and deployments remain.
+- Action: Required-gate failed because `crates/oasis7_net/src/libp2p_net/tests.rs` had grown to 1213 lines, exceeding the 1200-line test file limit. Moved the three `filter_request_peers_by_health` tests into `crates/oasis7_net/src/libp2p_net/tests/request_peer_filter_tests.rs` and registered the new test module from `tests.rs`.
+- Validation Command: `cargo fmt --check`; `./scripts/check-rust-file-size.sh`; `env -u RUSTC_WRAPPER cargo test -p oasis7_net filter_request_peers_by_health -- --nocapture`; `env -u RUSTC_WRAPPER cargo test -p oasis7_net libp2p_net::tests -- --nocapture`.
+- Expected Result: Rust formatting remains stable; the file-size gate reports zero oversized files; moved request filter tests and the full libp2p test module still pass.
+- Actual Result: Formatting passed. File-size gate passed with oversized code files=0 and test files=0. Request filter tests passed 3 tests. `oasis7_net` libp2p tests passed 57 tests.
+- Blocker / Next Action: Commit and push the test split, watch GitHub checks, then merge/package/deploy after checks and PR comments are clear.
